+++ frontend/src/pages/RegisterdFood.jsx
@@ -1,10 +1,18 @@
+import { FormProvider, useForm } from "react-hook-form";
 import { RegisterdFoodItem } from "../components/RegisterdFoodList/RegisterdFoodItem";
 import { ValidateErrorProvider } from "../context/ValidateErrorContext";
 
 export const RegisterdFood = () => {
+  const methods = useForm({
+    mode: "onBlur",
+    criteriaMode: "all"
+  });
+
   return (
-    <ValidateErrorProvider>
-      <RegisterdFoodItem />
-    </ValidateErrorProvider>
+    <FormProvider {...methods}>
+      <ValidateErrorProvider>
+        <RegisterdFoodItem />
+      </ValidateErrorProvider>
+    </FormProvider>
   );
 };
